Don't send validation errors to Sentry.
